Add ability to remove caption from PDF output
See https://github.com/LaurentRDC/pandoc-plot/issues/37.

For whatever reason, the change introduced in the PR related to the
issue above seems to not have an effect any more. Bring that change
back. Seems we might have problems with pandoc-crossref and other
filters because of which the change was dropped but right now this
is the only filter we care about.